another round of requested changes
