+++ src/frontend/src/pages/NewCalendarPage/CalendarDayCard.tsx
@@ -18,6 +18,7 @@ import DescriptionIcon from '@mui/icons-material/Description';
 import ArticleIcon from '@mui/icons-material/Article';
 import HelpIcon from '@mui/icons-material/Help';
 import EventPartialInfoView from './EventPartialInfoView';
+import { getConvertedEnd, getConvertedStart } from '../../utils/datetime.utils';
 
 export const getTeamTypeIcon = (teamTypeName: string, isLarge?: boolean) => {
   const teamIcons: Map<string, JSX.Element> = new Map([
@@ -37,34 +38,6 @@ export const getStatusIcon = (status: string, isLarge?: boolean) => {
     ['DONE', <CheckCircleIcon fontSize={isLarge ? 'large' : 'small'} />]
   ]);
   return statusIcons.get(status);
-};
-
-export const getConvertedStart = (event: Event, dayOfWeek: DayOfWeek) => {
-  const specificSlot = event.scheduledTimes.find((slot) => slot.days.find((day) => day === dayOfWeek));
-  const initialTime = specificSlot?.startTime ? new Date(specificSlot.startTime) : new Date(Date.now());
-  const startTime = new Date(
-    // getTimezoneOffset() is based on UTC (in minutes), and we need to inverse it and multiply by milliseconds to get the proper date
-    // for instance, if we were in UTC+3, it would return -180, which we need to invert and multiply to
-    initialTime.getTime()
-  );
-
-  const convertedStartTime = startTime.toLocaleTimeString('en-US', { hour: 'numeric', minute: '2-digit' });
-
-  return convertedStartTime;
-};
-
-export const getConvertedEnd = (event: Event, dayOfWeek: DayOfWeek) => {
-  const specificSlot = event.scheduledTimes.find((slot) => slot.days.find((day) => day === dayOfWeek));
-
-  const endTime = new Date(specificSlot?.endTime ?? Date.now());
-
-  const convertedEndTime = endTime.toLocaleTimeString('en-US', {
-    hour: 'numeric',
-    minute: '2-digit',
-    timeZoneName: 'short'
-  });
-
-  return convertedEndTime;
 };
 
 interface CalendarDayCardProps {
