+++ src/frontend/src/utils/datetime.utils.ts
@@ -1,4 +1,5 @@
 import dayjs from 'dayjs';
+import { DayOfWeek, Event } from 'shared';
 
 /**
  * Returns monday of current week
@@ -62,4 +63,32 @@ export const dateMonthDayYear = (date: Date): string => {
  */
 export const isPastEvent = (startDate: Date, endDate: Date) => {
   return startDate < endDate;
+};
+
+export const getConvertedStart = (event: Event, dayOfWeek: DayOfWeek) => {
+  const specificSlot = event.scheduledTimes.find((slot) => slot.days.find((day) => day === dayOfWeek));
+  const initialTime = specificSlot?.startTime ? new Date(specificSlot.startTime) : new Date(Date.now());
+  const startTime = new Date(
+    // getTimezoneOffset() is based on UTC (in minutes), and we need to inverse it and multiply by milliseconds to get the proper date
+    // for instance, if we were in UTC+3, it would return -180, which we need to invert and multiply to
+    initialTime.getTime()
+  );
+
+  const convertedStartTime = startTime.toLocaleTimeString('en-US', { hour: 'numeric', minute: '2-digit' });
+
+  return convertedStartTime;
+};
+
+export const getConvertedEnd = (event: Event, dayOfWeek: DayOfWeek) => {
+  const specificSlot = event.scheduledTimes.find((slot) => slot.days.find((day) => day === dayOfWeek));
+
+  const endTime = new Date(specificSlot?.endTime ?? Date.now());
+
+  const convertedEndTime = endTime.toLocaleTimeString('en-US', {
+    hour: 'numeric',
+    minute: '2-digit',
+    timeZoneName: 'short'
+  });
+
+  return convertedEndTime;
 };
